Rename datapoints -> stats
